+++ lib/art.dart
@@ -8,3 +8,4 @@ export 'src/widgets/art/exclamation_rectangle.dart';
 export 'src/widgets/art/lit_life_blurred_background_logo.dart';
 export 'src/widgets/art/lit_life_smiley_image.dart';
 export 'src/widgets/art/lit_lock_icon.dart';
+export 'src/widgets/art/lit_info_icon.dart';